Preview Enhancements (#14)
* Improve TR preview

Add buttons to view the match context and edit a topic, make the match details closeable, remove tab control and unused grids and embedded IE browser.

* Add tests, improve match UI

Add border, make the show matches button hide the keywords if clicked a second time, trim the final whitespace, etc.

* Add ribbon button to control match info pane

There's an X button to close it, there needs to be a way to get it back.  Now the user can click a checkbox to do so.